\name{*OrNULL-class}
\Rdversion{1.2}
\docType{class}
\alias{numericOrNULL-class}
\alias{vectorOrNULL-class}
\alias{matrixOrNULL-class}
\alias{characterOrNULL-class}
\alias{listOrNULL-class}
\alias{callOrNULL-class}


\title{Class *OrNULL}
\description{
class * or Null
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "*OrNULL" in the signature.
}
\examples{
showClass("callOrNULL")
}
\keyword{classes}
